+++ src/Exception/Error.php
@@ -9,4 +9,4 @@
 
 namespace Icicle\Stream\Exception;
 
-class Error extends \Error {}
+interface Error extends \Icicle\Exception\Error, Throwable {}