+++ src/MemorySink.php
@@ -9,8 +9,8 @@
 
 namespace Icicle\Stream;
 
+use Icicle\Exception\InvalidArgumentError;
 use Icicle\Stream\Exception\{
-    InvalidArgumentError,
     OutOfBoundsException,
     UnreadableException,
     UnseekableException,
@@ -23,7 +23,7 @@ use Icicle\Stream\Exception\{
  * may be determined with tell(). Since all data remains in the sink, the entire length of the sink is available with
  * getLength().
  */
-class MemorySink implements DuplexStreamInterface, SeekableStreamInterface
+class MemorySink implements DuplexStream, SeekableStream
 {
     /**
      * @var bool