+++ src/SeekableStream.php
@@ -9,7 +9,7 @@
 
 namespace Icicle\Stream;
 
-interface SeekableStreamInterface extends StreamInterface
+interface SeekableStream extends Stream
 {
     /**
      * @coroutine
@@ -25,12 +25,12 @@ interface SeekableStreamInterface extends StreamInterface
      *
      * @resolve int New pointer position.
      *
-     * @throws \Icicle\Stream\Exception\InvalidArgumentError If the whence value is invalid.
+     * @throws \Icicle\Awaitable\Exception\TimeoutException If the operation times out.
+     * @throws \Icicle\Exception\InvalidArgumentError If the whence value is invalid.
      * @throws \Icicle\Stream\Exception\InvalidOffsetException If the new offset would be outside the stream.
      * @throws \Icicle\Stream\Exception\UnseekableException If the stream is no longer seekable (due to being closed or
      *     for another reason).
      * @throws \Icicle\Stream\Exception\BusyError If the stream was already waiting on a read or seek operation.
-     * @throws \Icicle\Promise\Exception\TimeoutException If the operation times out.
      */
     public function seek(int $offset, int $whence = SEEK_SET, float $timeout = 0): \Generator;